Fixed major error where it treated doubles as ints, works much better now. Speed, friction and direction fully work!
git-svn-id: https://svn.code.sf.net/p/gjava/code@1166 0bf6c8e7-452b-0410-8c87-9aaedaf074f6

+++ G-Java/src/org/gcreator/compilers/gjava/api/Double.java
@@ -37,13 +37,13 @@ public class Double extends Object {
 
     @Override
     public Object sub(Object o) {
-        i-= (int) (o.getDouble());
+        i-= o.getDouble();
         return this;
     }
 
     @Override
     public Object add(Object o) {
-        i+= (int) (o.getDouble());
+        i+= o.getDouble();
         return this;
     }
 
@@ -73,13 +73,13 @@ public class Double extends Object {
 
     @Override
     public Object div(Object o) {
-        i /= (int) (o.getDouble());
+        i /= (o.getDouble());
         return this;
     }
 
     @Override
     public Object mult(Object o) {
-        i *= (int) (o.getDouble());
+        i *=  (o.getDouble());
         return this;
     }
